Move audio code to a YARG.Audio namespace
