+++ Assets/Script/GameManager.cs
@@ -3,6 +3,8 @@ using UnityEngine;
 using UnityEngine.Audio;
 using UnityEngine.InputSystem;
 using UnityEngine.SceneManagement;
+using YARG.Audio;
+using YARG.Audio.BASS;
 using YARG.Input;
 using YARG.Settings;
 using YARG.Song;
