+++ Assets/Script/UI/MusicLibrary/SongSelection.cs
@@ -8,6 +8,7 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.UI;
+using YARG.Audio;
 using YARG.Data;
 using YARG.Input;
 using YARG.Song;
