+++ Assets/Script/UI/MusicPlayer.cs
@@ -2,6 +2,7 @@ using Cysharp.Threading.Tasks;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
+using YARG.Audio;
 using YARG.Settings;
 using YARG.Song;
